Improve test VM workflow.
The VM that is started by `nix run .#start-vm` now uses GitHub
authentication, bringing it closer to the real thing. For this to work
it needs a GitHub client ID and secret, which it fetches from Vault.

On the host machine, before starting the VM, a vault token is obtained
using the `vault login` command. The token is passed to the VM as a
firmware parameter, allowing it to be used inside the VM to fetch the
secrets.

The file layout is tidied up a bit, and the VM tests are improved. These
still use basic authentication, for now at least. I've added a GitHub
Actions workflow to run the test.