Update Packit.
A bit of tweaks needed to expose the Git revision. By default Nix strips
out the `.git` folder. There are ways to keep it, but it tends to be
flaky. This replaces the metadata that would have been fetched from git
with what we have in the `sources.json`. Works similarly to how we
already do it in outpack_server.

The new metrics port needs to be defined, as otherwise all instances
were listening on the same port. Metrics are added to in the
metrics-proxy, and should get collected by Prometheus automatically.