Refactor secret fetching.
This replaces hard coded bash scripts with a more flexible Python script
and a list of secrets in the Nix configuration. This removes duplication
between the production script and the VM one, and will make it easier to
add new machines that may need a different list of secrets.

The actual flow to fetch secrets is unchanged. In production someone
needs to manually SSH onto the server and run `fetch-secrets`. The local
VM receives a vault token from the host and fetches the secrets
automatically.