Refactor and tidy the flake definition.
This removes a bunch of duplication in the deploy/start-vm/diff scripts
left over from the introduction of the wpia-packit-private machine.
These script now only exist as one script, that accepts a machine name
as their argument.

In order to split up the `flake.nix` file, it makes use of the
[`flake-parts`][flake-parts] library. This library uses the same module
system as NixOS itself, but at the flake configuration level itself.

The deploy script is changed from using `nixos-rebuild switch` to our
own implementation of it. The primary goal of this is to ensure
the derivation of the built system gets deployed too, allowing us to
reliable diff the system state in the future.

[flake-parts]: https://flake.parts/