Add a "training" Packit instance.
The instance is accessible to all members of the mrc-ide organisation.
The URL for it will be `https://packit.dide.ic.ac.uk/training`.

To make onboarding of new users as seamless as possible, all new users
are granted the `USER` role. Once the instance is deployed I will create
the role with general read/write permissions.