Add a GitHub workflow to compare system configuration.
This workflow runs `nix-diff` on both the target branch and on the PR
branch and posts the results as a comment to the PR.

The diff output isn't always the most friendly to look at in detail, but
it can help notice unintended changes. Note that this always compares
to the main branch, not to the live systems, since the latter wouldn't
be reachable from GitHub actions.