Automatically deploy from GitHub branches.
The machines each have a dedicated branch, named `deploy/<hostname>`.
They periodically pull the configuration from that branch and deploy it.

Separately, we will add a process for managing this branch using GitHub
action. A likely flow will be to synchronize the deploy branches to main
by default, but target it to feature branches when a label is applied to
a pull request.

For now this is only enabled for wpia-packit-dev. Once I get a bit more
experience with it and am happy with the workflow we will consider using
it on the production machines.