Remove a bit of duplication across machines.
A few bits were repeated across the two machine definitions, in
particular the Vault secrets configuration. We can remove that
duplication by moving the common parts into the `common/` directory.

The parts that differ by machine, ie. the path to the secret within
Vault, is left in the machine specific file. To make the split easier,
the list of secrets is turned into an attribute, allowing the same
secret to be referenced across files.