Simplify the churn caused by the static metrics.
We have a metric that reports the revision of the configuration, and its
hash. The latter causes a lot of churn any time any character is
changed, and probably does not add a lot of value. The Git revision is
useful enough on its own.

Additionally, the metric changes on every deployment, which was causing
the node exporter to restart. We can avoid that by adding a level of
indirection, by making the node exporter scan a directory in `/etc`, and
populate that directory instead. The exporter will pick up new versions
of the file without needing to restart.