Make the helper scripts hermetic.
We have a handful of scripts that operate on the NixOS host
configurations, for deploying / comparing / running in a VM.

These scripts take in a hostname and build the system configuration for
that host. Previously the scripts did this by invoking
`nix build .#nixosConfiguration.<hostname>....` (or equivalent),
assuming that the command was run from within the repository.

This worked well if the user ran `nix run .#deploy <hostname>`, but
doesn't work so well if the user instead uses an arbitrary flake
reference to run the script, such as with
`nix run github:reside-ic/packit-infra#deploy <hostname>`. Instead of
using the system configuration from GitHub, this command would look for
it in the current working directory.

To make this pattern work better, the scripts are modified to embed a
path to the pre-evaluated system configurations. With this data made
available to the helpers scripts, they can run hermetically without
needing to evaluate the flake a second time.

The system configurations are only pre-evaluated, not pre-built. This
distinction is important since we don't want to have to build the
configuration for all hosts just to deploy a single one.