Allow running reports from GitHub actions on packit/reside.
This adds a new policy for OIDC authentication, allowing GitHub actions
on the `reside-ic/packit-infra-test-repo` repository to login to the
`packit/reside` instance and read/write from outpack, as well as invoke
the runner.

This has to be done on `wpia-packit` and not `wpia-packit-dev` as the
latter isn't exposed publicly.