Fix the enumeration of hostnames.
The last change to the deploy scripts broke the enumeration of available
hostnames, which gets printed when someone runs `nix run .#deploy`
without arguments.

I've also updated the command that gets commented on automatic update
pull requests to include the `--refresh` argument. Without it, Nix tends
to cache the repository contents, and might not end up deploying the
lastest version of it.