Add a runner-cli command.
This command starts an R session that is setup to interact with the rrq
queue used by the orderly runner. This can be used to debug or inspect
the state of the queue.

At startup, the R session attaches the `rrq` package and configures a
default controller using the same queue name as the orderly runner.

```sh
$ runner-cli
> rrq_worker_list()
[1] "dermatoplastic_coral"
```

Since `runner-cli` is just a wrapper around the R command, the usual
flags can be used to immediately evaluate an expression.

```
$ runner-cli -s -e 'rrq_worker_list()'
[1] "dermatoplastic_coral"
```

The R session actually runs inside a podman container, using the same
image as used by runner API and workers.