Use cache-nix-action in the test workflow.
The test workflow takes a significant amount of time to run (about 15
minutes) as it needs to rebuild outpack_server and Packit each time.

In theory Nix should be great as caching those built artefacts while
precisely be able to determine when they need rebuilding. This is more
or less what we had in place when we used the magic-nix-cache action.

Unfortunately that action was using an undocumented GitHub API, which
has now been sunset without a supported replacement.

The cache-nix-action action can be used instead, as it only uses the
documented GitHub APIs for this. However it works in a different way, as
it saves and restores the entire `/nix` store, whereas the
magic-nix-cache action would save and restore each built derivation as a
separate artefact.

Saving the whole store is not as powerful, and in particular there's a
risk of the store growing everytime we need to build a new versions of
the packages, until we hit GitHub's storage limit. Because the whole
store is a single artefact, GitHub's cache cannot evict the individual
packages we have not used in a long time. If and when this becomes an
issue we'll need to look into adding own our garbage collection stage at
the end of the workflow, where we only keep the derivations that we've
accessed.