Make packit-api a `notify` service.
Packit is pretty slow to start. Using `notify` as the service type
allows us to track within systemd whether it is still starting or if it
is started and accepting connections.

In the future this may even allow us to run some provisioning steps (eg.
creating basic users, creating roles, seeding the `reside-dev` instance
with test data, ...) that express a dependency on Packit being up and
running.