Fix usage of Nix cache in GitHub actions.
Turns out I'd misinterpreted the semantics of the primary key. If we use
a static key (as I had), the action will never overwrite an existing
cache, meaning we were just using the first cached store that was
created.

Instead we should use a cache key that reflects the contents of the
repository. If this key does not exist yet, the action will pull any
cached store that matches a common prefix.

In a more traditional setting, the cache key might depend on just the
contents of one file (say a `package-lock.json` if caching NPM
dependencies), but in this context we are interested in caching the
entire build so I think it makes sense to use the commit hash as the
cache key.