Make the configuration drift metrics more precise.
We have some alerts configured that are used to detect configuration
drift between the state of the repository and what is actually deployed
on the machines.

Currently the alerts fire if the Git revision between the two differs,
which makes the alerts very sensitive. Any change at all requires all
machines to be redeployed, even if the change only affects a subset of
them (or none, in the case of a documentation-only change for example).

A better way to detect this is to track the hash of the built system
instead, which Nix makes very easy to obtain.

In the past we could get Prometheus to scrape the revision of the
repository by calling the GitHub API. To get the hash of the built
system requires a little more complexity, as we need to actually build
the repository. Rather than build repeatedly on every scrape, we can use
GitHub actions to produce the metrics and host them onto GitHub pages.
We can point Prometheus directly at the right URL and get metrics about
the build.

There are a few more caveats to including the system's hash in the
output. These are described in detail, along with the workaround used,
in `machines/common/configuration-info.nix`.