Get certificates for packit-dev from Let's Encrypt.
To get the certificates we need to prove ownership of the domains in an
automated fashion. The easiest way to do this is by hosting a challenge
on an http server and have Let's Encrypt fetch that challenge and check
its contents.

Unfortunately some of our machines are behind a firewall and are only
accessible from within the DIDE network. To prove ownership of these
domains, we have to use a DNS challenge instead. This requires us to be
able to programatically configure a TXT DNS record on
`_acme-challenge.<hostname>.dide.ic.ac.uk`. Our DNS entries are managed
by ICT, using "HDB" as their source of truth. ICT provides an API to add
entries for this exact purpose.

ICT did give us a certbot plugin to interact with their API, however
NixOS already has very good integration with a different ACME client,
Lego. Adding a new DNS provider implementation to Lego was quite easy,
as we can just get Lego to call out to an external script, and all that
script needs to do is make a `PUT` or `DELETE` call to the HDB API.

For now only packit-dev is using Let's Encrypt, although there is
nothing preventing us from using it on other hosts - I will probably
deploy it to the others very soon.

The local VMs never use Let's Encrypt. They use the same self-signed
certificates we were already using before.