Update all, add script to update orderly.runner.
The script is a basic wrapper around nix-prefetch-docker.

The way images are referenced is simplified. We used to have a two step
process to start images, first loading the image tarball into podman and
then starting the container. It turns out we can just use
`docker-archive:/path/to/image.tar` instead. Podman will take care of
importing the image as needed. This removes the oddity surrounding the
`finalImageTag` that the previous version had.