Use the vimc-robot account to create the pull request.
Pull request created by an action's default GITHUB_TOKEN do not trigger
any further action runs (presumably to avoid cycles). This was affecting
the nightly update action, and required manual intervention to get the
CI to run.

By using a token that is tied to a specific account we can workaround
this. I have created a fork of this repository under
vimc-robot/packit-infra and added a PAT that can push to it and create
pull requests.

The PAT has scopes `repo` and `workflow`. The latter is needed in case
the workflow files on the upstream repository change and need to be
synced onto the forked branch. I tried using fine-grained token for this
but it did not seem to work as there is not way to get a fine-grained
token for a repo on an org that you are not a part of.

Additionally the vimc-robot account was added as a collaborator to this
repo with "Triage" permissions, allowing it to add reviewers to its pull
requests.